+++ sql/update_to_1.8.0.sql
@@ -1,0 +1,5 @@
+
+CREATE TABLE theme (
+    name VARCHAR (256) PRIMARY KEY,
+    content TEXT
+);