+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Theme endpoints added
+- Tests for themes
+- SQL migration script
+
+### Changed
+- SQL script for eva databases (theme table)
 
 ## [1.7.1] - 2021-07-13
 ### Added